backend/requirements.txt
Django==3.1
uWSGI==2.0.19
django-environ==0.4.5
psycopg2-binary==2.8.6
djangorestframework-gis==0.16
django-cors-headers==3.5.0
